Added a new `Data Analysis Simple Agent with PydanticAI` notebook
This notebook is based on the LangChain `Data Analysis Simple Agent` notebook and demonstrates
how we can do the same thing with PydanticAI.

It demonstrates tool creation with PydanticAI, retries and other mechanisms. It also shows that
unlike LangChain who has many built-in tools, the PydanticAI echo-system is still young, so we may
need to implement some of the functionality that LangChain alreaady has, ourselves.

Finally, improved some of the wording in the original PydanticAI notebook
`Building a Conversational Agent with Context Awareness`